create popup for remove task

+++ my-app/src/component/Task/Tasks.js
@@ -4,6 +4,8 @@ import EditableLabel from 'react-inline-editing';
 import {Button, Form, FormGroup, Label, Input, Card, CardHeader, CardBody, Container, Col, Table} from 'reactstrap';
 import InlineEditable from "react-inline-editable-field";
 import {browserHistory} from 'react-router';
+import {confirmAlert} from 'react-confirm-alert'; // Import
+import 'react-confirm-alert/src/react-confirm-alert.css' // Import css
 
 
 class Tasks extends Component {
@@ -108,16 +110,28 @@ class Tasks extends Component {
   };
 
   handleRemoveSpecificRow = (i, item) => () => {
-    const rows = [...this.state.tasks];
-    rows.splice(i, 1);
-    this.setState({tasks: rows});
-    fetch(`/api/tasks/${item.id}`,
-      {
-        method: "DELETE",
-        headers: new Headers({
-          'Authorization': `Bearer ${localStorage.getItem('token')}`
-        }),
-      }).then(response => response.json())
+    confirmAlert({
+      title: 'Remove this task',
+      message: 'Are you sure?',
+      buttons: [
+        {
+          label: 'Yes',
+          onClick: () => {
+            const rows = [...this.state.tasks];
+            rows.splice(i, 1);
+            this.setState({tasks: rows});
+            fetch(`/api/tasks/${item.id}`,
+              {
+                method: "DELETE",
+                headers: new Headers({'Authorization': `Bearer ${localStorage.getItem('token')}`}),
+              }).then(response => response.json())
+          }
+        },
+        {
+          label: 'No'
+        }
+      ]
+    })
   };
 
   toggleChange = () => {
@@ -183,6 +197,7 @@ class Tasks extends Component {
       }).then((response) => response.json()).then((data) => this.setState({tasks: data}))
   }
 
+
   render() {
     return (
       <div>
@@ -197,9 +212,7 @@ class Tasks extends Component {
           <Button className='mx-3' color="warning" onClick={this.removeCheckTasks}>Remove all</Button>
         </div>
 
-
         <div className='d-flex w-100'>
-
           <div className='w-75 mx-5'>
             <h3>Active</h3>
             <Table hover={true} className='my-5'>
@@ -254,7 +267,8 @@ class Tasks extends Component {
                     }}/></td>
                     <td className="description">{item.priority}</td>
                     <td className="description">{item.date_task}</td>
-                    <td><Input type='checkbox' id={item.id} checked='checked' name="status" onChange={this.changeTask(item)}/></td>
+                    <td><Input type='checkbox' id={item.id} checked='checked' name="status"
+                               onChange={this.changeTask(item)}/></td>
                     <td><Button color="danger" id={item.id}
                                 onClick={this.handleRemoveSpecificRow(i, item)}>Remove</Button></td>
                   </tr> : false
